requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.0
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
phonenumbers==8.12.54
Pillow==9.2.0
sqlparse==0.4.2